chore: moving navigationwrapper to allow for use of hook
